fix(audit): host test build, stale calibration metadata, and DDSM model comment
- pi_link.c: fix GCC sign-compare warning promoted to error by -Werror

- tests/CMakeLists.txt: link libm for kinematics math functions

- tests/test_ddsm315.c: update velocity expectation after 0.1 RPM unit fix

- rl/env/fivebar_ik_table.json: update embedded model_hash to current 9838e7fe323c1809

- stm32_firmware/Core/Src/ddsm315.c: correct DDSM115 -> DDSM315 in comment

+++ stm32_firmware/Core/Src/pi_link.c
@@ -150,7 +150,7 @@ int pi_link_parse_packet(const uint8_t *buf, uint16_t len) {
 
     int parsed = 0;
     uint16_t offset = 0;
-    while (rx_stream_len - offset >= 12u) {
+    while ((uint32_t)(rx_stream_len - offset) >= 12u) {
         if (rx_stream[offset] != PI_FRAME_HEADER) {
             ++offset;
             continue;
